feat(dispatcher/routing): core distance related classes added/moved
